Initial Commit!
Currently, we can send messages to a hardcoded Discord user, and we
listen for all events that the bot is able to see. When the bot recieves
a message, it currently just prints it to the console, but we wanna
make a request to a webhook in a later version.