Parameterise messaging
Now we can send to specific users and channels if we know their IDs.
There are two endpoints for this. Still need to figure out if we need to
support multiple bot users (probably), and deal with incoming messages
in a more elegant way than just printing them. If we need to have more
than just one bot user, we'll also need some kind of
authentication/identification of which bot to send from.