Refactor config to separate file with Pydantic
We want to eventually be able to have multiple bots with our proxy. This
is the first step to that. Ultimately, I expect the config will be
coming straight from the postgres database that RP uses, but we'll see.
Until then, we can use the config.toml. We don't currently support
multiple bots, but that will come soon.