Update dockerfile with CMD, add toml, env vars
Add a default command to the dockerfile, in this case, uvicorn, so that
things work nicely with compose. Also, everything was broken because
toml wasn't actually in the dependencies. TBH we probably don't need it,
as we pull the config from the db now, but that's a later refactor. Pull
the db connection string and rapidpro root domain from the environment,
so as to work with compose.