:lipstick: Fix lint errors, remove call to parse_config_file
This commit fixes some of the more glaring linting errors, adding
docstrings and whatnot to most things. Also added bandit to the deps,
which helps catch security errors. One thing we do need to ensure is
that we verify https in the case that we aren't dealing with a local
url in our attachments list. That's an important todo.

Also removed the unnecessary call to parse_config_file.