WIP Add more roles stuff
We need to use non-default intents to see what is going on with our
roles. Also forgot to send user info up with the role change event. TODO
add a config for the token we are using to talk to django, as it's
hardcoded right now (not a real secret)